requirements.txt
streamlit>=1.32.0
google-auth-oauthlib>=1.2.0
google-ads>=26.0.0
openai==1.12.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
umap-learn>=0.5.5
python-dotenv>=1.0.0
plotly>=5.18.0
requests>=2.31.0
sentence-transformers>=2.5.0